+++ src/domible/builders/__init__.py
@@ -3,6 +3,6 @@
 from domible.builders.tableBuilder import RowBuilder, TableBuilder, buildTableFromDicts
 from domible.builders.navBuilder import NavBuilder 
 from domible.builders.listBuilder import ListBuilder 
-
+from domible.builders.elementFromObject import element_from_object  
 
 # end of file